% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit.R
\name{AIC.breathtestnlmefit}
\alias{AIC.breathtestnlmefit}
\title{S3 AIC method for breathtestnlmefit}
\usage{
\method{AIC}{breathtestnlmefit}(object, ...)
}
\arguments{
\item{object}{of class breathtestnlmefit}

\item{...}{not used}
}
\description{
Extract AIC from a model fitted with \code{\link{nlme_fit}}
}
