% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breathtestcore-package.R
\name{breathtestcore-package}
\alias{breathtestcore-package}
\alias{breathtestcore}
\title{breathtestcore}
\description{
Stan fits to 13C breath test curves. Reads several formats
  of 13C   data (IRIS/Wagner, BreathID) and CSV.  Creates artificial
  sample data for testing.  Fits Maes/Ghoos, Bluck-Coward self-correcting
  formula using 'nls', 'nlme'. Methods to fit breath test curves with
  Bayesian Stan methods are refactored to package 'breathteststan'.
  For a Shiny GUI, see package 'dmenne/breathtestshiny' on github.
}
