% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_iris_csv.R
\name{extract_id}
\alias{extract_id}
\title{Extracts an ID from string IRIS CSV file}
\usage{
extract_id(id)
}
\arguments{
\item{id}{One item from column Identifikation, e.g. "KEK-ZH-Nr.2013-1234"}
}
\description{
First tries to extract only digits, separating these by underscore
when there are multiple blocks. If this give a non-valid  id, returns the
whole string without spaces and periods, hoping it makes sense.
For internal use, but should be overridden for exotic IDs
}
\examples{
extract_id
}
