% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_fit.R
\name{null_fit}
\alias{null_fit}
\title{Convert data to class breathtestfit}
\usage{
null_fit(data, ...)
}
\arguments{
\item{data}{Data frame or tibble as created by \code{\link{cleanup_data}},
with mandatory columns \code{patient_id, group, minute} and \code{pdr}.}

\item{...}{Not used}
}
\value{
A list of classes \code{breathtestnullfit, breathtestfit}
with element \code{data} which contains the unmodified data.
}
\description{
Does not change the data set, but returns a class suitable
for plotting raw data with \code{\link{plot.breathtestfit}}.
See \code{\link{read_any_breathtest}} for an example.
}
