% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_breathid_xml.R
\name{read_breathid_xml}
\alias{read_breathid_xml}
\title{Read new BreathID/Examens XML file}
\usage{
read_breathid_xml(filename = NULL, text = NULL)
}
\arguments{
\item{filename}{name of xml-file to be read}

\item{text}{alternatively, text can be given as string}
}
\value{
List of structure of class \code{\link{breathtest_data}}; an XML file
can contain multiple data sets.
}
\description{
Reads 13c data from an XML BreathID file, and returns a stucture 
of class \code{breathtest_data}.
}
\examples{
filename = system.file("extdata", "NewBreathID_01.xml", package = "breathtestcore")
# Show first lines
cat(readLines(filename, n = 10), sep="\\n")
#
bid = read_breathid_xml(filename)
str(bid)
}
