% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_fit.R
\name{sigma_fit}
\alias{sigma_fit}
\title{S3 method to exctract the residual error}
\usage{
sigma_fit(fit)
}
\value{
A numeric value giving the average residual error. For nls fit, 
the mean of the deviances is returned.
}
\description{
Methods for nls, nlme are available; additional functions
for Stan-based fits are defined in package `breathteststan`.
}
