% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usz_13_sol_liq.R
\name{usz_13c_sol_liq}
\alias{usz_13c_sol_liq}
\title{Subset of usz_13c with solid/liquid meals}
\usage{
usz_13c_sol_liq(sample_minutes = 15)
}
\arguments{
\item{sample_minutes}{When median inter-sample period is less than this value,
raw data are resampled to the given period. Resampling is required, because
some methods do not converge when there are to many serially correlated data.}
}
\value{
A `tibble` of records from patients and normals with two meals,
possibly resampled, with columns `patient_id`, `group` (`liquid_normal, solid_normal,
solid_patient, liquid_patient`) and `pdr`
}
\description{
Get subset of clinical data with records for solid and liquid meals of the same patient
}
\examples{
  fit = usz_13c_sol_liq() |>
    nlme_fit()
  coef(fit) |>
    dplyr::filter(parameter == "t50", method == "maes_ghoos")
  fit |>
    plot()
}
