% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breathtestfit_broom.R
\name{augment.breathtestfit}
\alias{augment.breathtestfit}
\title{Augmented prediction for breathtest fit}
\usage{
\method{augment}{breathtestfit}(x, by = NULL, dose = 100, ...)
}
\arguments{
\item{x}{Object of class \code{breathttestfit}}

\item{spacing}{if given, time spacing of}
}
\value{
\code{augment} When \code{by} is NULL,  returns one row for each 
original observation, and column \code{fitted}.
When \code{by} is a positive number, it is used as a step size for
a sequence of minutes from 0 to the maximum value of minute in data set. 
One row for each new time value with is returned, with column \code{fitted}.
}
\description{
Broom-methods to compute predicted values from  the results 
of class \code{breathttestfit}  as generated by  \code{breathtestcore::nls_fit} 
or \code{breathtestcore::nlme_fit}.
}
