% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breathtestfit_broom.R
\name{augment.breathtestfit}
\alias{augment.breathtestfit}
\title{Augmented prediction for breathtest fit}
\usage{
\method{augment}{breathtestfit}(x, by = NULL, minute = NULL, dose = 100,
  ...)
}
\arguments{
\item{x}{Object of class \code{breathttestfit}}

\item{by}{When \code{by} is NULL, predictions for the original data values
are returned.  When \code{by} is a positive number, it is used as a step 
size for a sequence of minutes from 0 to the maximum value of minute in data set.}

\item{minute}{When a vector is passed, this overrides settings in \code{by},
and predictions are calculated at the requested minute values.}

\item{dose}{13C acetate or octanoate dose}

\item{...}{other parameters passed to methods}
}
\value{
\code{augment} When \code{by} is NULL,  returns one row for each 
original observation, and column \code{fitted}.
}
\description{
Broom-methods to compute predicted values from  the results 
of class \code{breathttestfit}  as generated by  \code{breathtestcore::nls_fit} 
or \code{breathtestcore::nlme_fit}.
}
\examples{
data = cleanup_data(simulate_breathtest_data()$data)
fit = nls_fit(data)
augment(fit, minute = c(60, 120))
}
