% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breath_test_read_function.R
\name{breathtest_read_function}
\alias{breathtest_read_function}
\title{Snoop method to read breath test file}
\usage{
breathtest_read_function(filename = NULL, text = NULL)
}
\arguments{
\item{filename}{breath test data file from Iris/Wagner, BreathID or generic CSV,}

\item{text}{as alternative to filename, the text may be given as string}
}
\value{
function to read the file or the text; NULL if no matching function 
was found
}
\description{
Reads the first line of a file, and returns
the best matching function to read the breath test data in it.
}
