% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btcore_file.R
\name{btcore_file}
\alias{btcore_file}
\title{Path to example breath test data file}
\usage{
btcore_file(filename = NULL, full.names = FALSE)
}
\arguments{
\item{filename}{example file in \code{extdata} directory without path. 
Case sensitive on Unix. When filename is missing, returns the names of the 
available sample files.}

\item{full.names}{When \code{filename} is NULL, return full path names}
}
\value{
full filename to example file to use in read_xxx
}
\description{
Path to example breath test data file
}
\examples{
  head(btcore_file())
  filename = btcore_file("IrisMulti.TXT")
  data = read_iris(filename)
}
