% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btcore_file.R
\name{btcore_file}
\alias{btcore_file}
\title{Path to example breath test data file}
\usage{
btcore_file(filename)
}
\arguments{
\item{filename}{example file in \code{extdata} directory without path. 
Case sensitive on Unix.}
}
\value{
full filename to example file to use in read_xxx
}
\description{
Path to example breath test data file
}
\examples{
  filename = btcore_file("IrisMulti.TXT")
  data = read_iris(filename)
}
