% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_data.R
\name{cleanup_data}
\alias{cleanup_data}
\title{Transforms 13C breath data into a clean format for fitting}
\usage{
cleanup_data(data)
}
\arguments{
\item{data}{A data frame with at least two numeric columns with names \code{minute} 
and \code{pdr} which can be used to fit a single 13C record}
}
\value{
A dat
}
\description{
Accepts various data formats of ungrouped or grouped 13C breath 
test time series, and transforms these into a data frame that can be used by
all fitting functions
}
