% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_data.R
\name{cleanup_data}
\alias{cleanup_data}
\title{Transforms 13C breath data into a clean format for fitting}
\usage{
cleanup_data(data)
}
\arguments{
\item{data}{\itemize{
  \item{A data frame or tibble with at least two numeric columns with 
names \code{minute} and \code{pdr} which can be used to fit a single 13C record }
   \item{A list of data frames/tibbles that are concatenated}
}}
}
\value{
A tibble with 2, 3 or 4 columns. patient_id and group are coerced to 
character.
}
\description{
Accepts various data formats of ungrouped or grouped 13C breath 
test time series, and transforms these into a data frame that can be used by
all fitting functions.
}
