% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_fit.R
\name{coef.breathtestfit}
\alias{coef.breathtestfit}
\title{S3 coef for breathtestfit}
\usage{
\method{coef}{breathtestfit}(object, ...)
}
\arguments{
\item{object}{of class breathtestfit, as returned by nls_fit or nlme_fit}

\item{...}{other parameters passed to methods}
}
\description{
Extracts the coefficients part from fitted 13C beta exponential
model; same as \code{fit$coef}
}
