% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_beta.R
\name{cum_exp_beta}
\alias{cum_exp_beta}
\title{Cumulative exponential beta function}
\usage{
cum_exp_beta(minute, dose, cf)
}
\arguments{
\item{minute}{time in minutes}

\item{dose}{in mg}

\item{cf}{named vector of coefficients; only \code{k} and \code{beta} are required.
Note that \code{k} is measured in 1/min (e_g_ 0_01/min),
while often it is quoted as 1/h (e_g_ 0_6/h).}
}
\value{
Vector of predicted cumulative pdr
}
\description{
Equation (2), page 4 from Bluck, "Recent advances in the interpretation of
the 13C octanoate breath test for gastric emptying"
}
\seealso{
\code{\link{exp_beta}}
}
