% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit.R
\name{nlme_fit}
\alias{nlme_fit}
\title{Mixed-model nlme fit to 13C Breath Data}
\usage{
nlme_fit(data, dose = 100, start = list(m = 30, k = 1/100, beta = 2),
  sample_minutes = 15)
}
\arguments{
\item{data}{Data frame or tibble as created by \code{\link{cleanup_data}},
with mandatory columns \code{patient_id, group, minute} and \code{pdr}.
It is recommended to run all data through \code{\link{cleanup_data}} which
will insert dummy columns for \code{patient_id} and \code{minute} if the
data are distinct, and report an error if not. At least 3 records are required
for a population fit, but 10 or more are recommended to obtain a stable result.}

\item{dose}{Dose of acetate or octanoate. Currently, only one common dose
for all records is supported.}

\item{start}{Optional start values.}

\item{sample_minutes}{If mean sampling interval is < sampleMinutes, data are subsampled
using a spline algorithm}
}
\value{
A list of class "breathtestfit" with elements
\itemize{
  \item {\code{coef} Estimated parameters as data frame in a key-value format with
   columns \code{patient_id, group, parameter, method} and \code{value}.
   Has an attribute AIC.}
   \item {\code{data}  The effectively analyzed data. If density of points
   is too high, e.g. with BreathId devices, data are subsampled before fitting.}
}
}
\description{
Fits exponential beta curves to 13C breath test series data using
a mixed-model population approach. See
\url{https://menne-biomed.de/blog/breath-test-stan} for a comparision between
single curve, mixed-model population and Bayesian methods.
}
\examples{
d = simulate_breathtest_data(n_records = 3, noise = 0.2, seed = 4711)
data = cleanup_data(d$data)
fit = nlme_fit(data)
plot(fit) # calls plot.breathtestfit
options(digits = 2)
library(dplyr)
cf = coef(fit)
cf \%>\%
  filter(grepl("m|k|beta", parameter )) \%>\%
  select(-method, -group) \%>\%
  tidyr::spread(parameter, value) \%>\%
  inner_join(d$record, by = "patient_id") \%>\%
  select(patient_id, m_in = m.y, m_out = m.x,
         beta_in = beta.y, beta_out = beta.x,
         k_in = k.y, k_out = k.x)
}
\seealso{
Base methods \code{coef, plot, print}; methods from package
 \code{broom: tidy, augment}.
}
