% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_fit.R
\name{nls_fit}
\alias{nls_fit}
\title{Single curve fit with nls to 13C breath test data}
\usage{
nls_fit(data, dose = 100, start = list(m = 50, k = 1/100, beta = 2))
}
\arguments{
\item{data}{Data frame or tibble as created by \code{\link{cleanup_data}}, 
with mandatory columns \code{patient_id, group, minute} and \code{pdr}. 
It is recommended to run all data through \code{\link{cleanup_data}} which
will insert dummy columns for \code{patient_id} and \code{minute} if the
data are distinct, and report an error if not.}

\item{dose}{Dose of acetate or octanoate. Currently, only one common dose
for all records is supported.}

\item{start}{Optional start values
\code{patient_id} and \code{group}.}
}
\value{
Estimated parameters are returned in a key-value format with 
columns \code{patient_id, group, parameter, method} and \code{value}
}
\description{
Single curve fit with nls to 13C breath test data
}
