% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_breathtestfit.R
\name{plot.breathtestfit}
\alias{plot.breathtestfit}
\title{S3 plot method for breathtestfit}
\usage{
\method{plot}{breathtestfit}(x, inc = 5, method_t50 = "maes_ghoos", ...)
}
\arguments{
\item{x}{object of class \code{breathtestfit}, as returned by \code{nls_fit}, 
\code{nlme_fit}  or \code{stan_fit} from package \code{breathteststan}}

\item{inc}{increment for fitted curve plot in minutes}

\item{method_t50}{method for t50: \code{maes_ghoos}, \code{bluck_coward} or 
\code{maes_ghoos_scint}}

\item{...}{other parameters passed to methods}
}
\description{
Plots 13C data and fits.
}
\details{
A description of plotting
}
\examples{
data = list(
  A = simulate_breathtest_data(n_records = 6, seed = 100)$data,
  B = simulate_breathtest_data(n_records = 4, seed = 187)$data 
)
# cleanup_data combines the list into a data frame
x = nls_fit(cleanup_data(data))
plot(x)
}
