% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_breathtestfit.R
\name{plot.breathtestfit}
\alias{plot.breathtestfit}
\title{S3 plot method for breathtestfit}
\usage{
\method{plot}{breathtestfit}(x, ...)
}
\arguments{
\item{x}{object of class breathtestfit, as returned by nls_fit or nlme_fit}

\item{...}{other parameters passed to methods}
}
\description{
Plots 13C data and fits.
}
