% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_breathid.R
\name{read_breathid}
\alias{read_breathid}
\title{read BreathID file}
\usage{
read_breathid(filename)
}
\arguments{
\item{filename}{name of txt-file to be read}
}
\value{
structure of class \code{\link{breathtest_data}}
}
\description{
Reads 13c data from a BreathID file, and returns a stucture for
of class \code{breathtest_data}. BreathID files start with the following lines

\code{Test and Patient parameters}
}
\examples{
filename = system.file("extdata", "350_20043_0_GER.txt", package = "breathtestcore")
bid = read_breathid(filename)
str(bid)
}
\author{
dieter menne, \email{dieter_menne@menne-biomed_de}
}
