% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_breathid.R
\name{read_breathid}
\alias{read_breathid}
\title{Read BreathID file}
\usage{
read_breathid(filename = NULL, text = NULL)
}
\arguments{
\item{filename}{name of txt-file to be read}

\item{text}{alternatively, text can be given as string}
}
\value{
Structure of class \code{\link{breathtest_data}}
}
\description{
Reads 13c data from a BreathID file, and returns a stucture 
of class \code{breathtest_data}.
}
\examples{
filename = system.file("extdata", "350_20043_0_GER.txt", package = "breathtestcore")
# Show first lines
cat(readLines(filename, n = 10), sep="\\n")
#
bid = read_breathid(filename)
str(bid)
}
