% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_iris.R
\name{read_iris}
\alias{read_iris}
\title{Read 13C data from IRIS/Wagner Analysen}
\usage{
read_iris(filename)
}
\arguments{
\item{filename}{name of iris/wagner file in composite format}
}
\value{
list with \code{file_name, patient_name, patient_first_name,
test, identifikation}, and data frame \code{data} with \code{time}
and \code{dob}
}
\description{
Reads composite files with 13C data from IRIS/Wagner Analysen.
The composite files start as follows:
\preformatted{
"Testergebnis"
"Nummer","1330"
"Datum","10.10.2013"
"Testart"}
}
\examples{
filename = system.file("extdata", "IrisMulti.txt", package = "breathtestcore")
iris_data = read_iris(filename)
str(iris_data)
}
\author{
dieter menne, \email{dieter_menne@menne-biomed_de}
}
