% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_iris_csv.R
\name{read_iris_csv}
\alias{read_iris_csv}
\title{Read 13C data from IRIS/Wagner Analysen in CSV Format}
\usage{
read_iris_csv(filename)
}
\arguments{
\item{filename}{Name of IRIS/Wagner file in CSV format}
}
\value{
list of class \code{breath_test_data} with file name, 
patient name, patient first name, test, identifikation,
and data frame \code{data} with \code{time} and \code{dob}
}
\description{
Reads 13C data from IRIS/Wagner Analysen in CSV Format
The CSV files start as follows:
\preformatted{
"Name","Vorname","Test","Identifikation"
} 
This format does not have information about the substrate (acetate, octanoate),
the dose and body weight and height. By default, using acetate as substrate, 
dose = 100,  weight = 75, height = 180.
}
\examples{
filename = system.file("extdata", "IrisCSV.TXT", package = "breathtestcore")
iris_data = read_iris_csv(filename)
str(iris_data)
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
