% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma.breathtestnlmefit}
\alias{sigma.breathtestnlmefit}
\title{S3 method to exctract the fit's residual standard deviation}
\usage{
\method{sigma}{breathtestnlmefit}(object, ...)
}
\arguments{
\item{object}{Result of class \code{breathtestfit}}

\item{...}{Not used}
}
\value{
A numeric value giving the standard deviation of the residuals.
}
\description{
Functions for \code{nls} and \code{nlme} are available; 
additional functions for Stan-based fits are defined in package \code{breathteststan}.
}
