% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_fit.R
\name{sigma_fit}
\alias{sigma_fit}
\title{S3 method to exctract the residual error}
\usage{
sigma_fit(fit)
}
\arguments{
\item{fit}{Result of class \code{breathtestfit}}
}
\value{
A numeric value giving the standard deviation of the residuals.
}
\description{
Functions for \code{nls} and \code{nlme} are available; 
additional functions for Stan-based fits are defined in package \code{breathteststan}.
}
