% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_data.R
\name{subsample_data}
\alias{subsample_data}
\title{Decimate densely sampled 13C time series}
\usage{
subsample_data(data, sample_minutes)
}
\arguments{
\item{data}{Data frame with columns \code{patient_id, group, minute, pdr}.}

\item{sample_minutes}{Required average density. When points are more closely
space, data are subsamples. No upsampling occurs when data are more sparse.}
}
\description{
When data of a record are more closely spaced \code{sample_minutes}, 
these are spline-subsampled to \code{sample_minutes}. Values up to 1/5 of 
maximal time are sampled more densely.
Too dense sampling leads to non-convergent \code{nlme} fits and to long runs
with Stan-based fits. 
The function is used internally by package \code{breathtestcore} and is exported 
for use  by package \code{breathteststan}.
}
