% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breathtest_parameters.R
\name{t50_maes_ghoos_scintigraphy}
\alias{t50_maes_ghoos_scintigraphy}
\title{t50 from maes with scintigrapic correction}
\usage{
t50_maes_ghoos_scintigraphy(cf)
}
\arguments{
\item{cf}{named vector of coefficients; only \code{k} and \code{beta} are required}
}
\value{
time where value is 1/2 of maximum, i.e. t50 in minutes.
}
\description{
t50 from beta exponential function, with linear correction for
scintigraphic values. This is for comparison with published data only;
there is little justification for using it, even if it is closer to real gastric
emptying times as determined by MRI or scintigraphy.
}
\seealso{
\code{\link{exp_beta}}, and \code{\link{t50_bluck_coward}} for an example.
}
