% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breathtest_parameters.R
\name{t50_maes_ghoos_scintigraphy}
\alias{t50_maes_ghoos_scintigraphy}
\title{Half-emptying time t50 from Maes/Ghoos fit with scintigraphic correction}
\usage{
t50_maes_ghoos_scintigraphy(cf)
}
\arguments{
\item{cf}{named vector of coefficients; only \code{k} and \code{beta} are required}
}
\value{
Time where value is 1/2 of maximum, i.e. t50 in minutes.
}
\description{
Half-emptying time t50 in minutes from beta exponential function fit,
with linear and rather arbitrary correction for 
scintigraphic values. This is given for comparison with published data only;
there is little justification to use it, even if it is closer to real gastric
emptying times as determined by MRI or scintigraphy.
}
\seealso{
\code{\link{exp_beta}}, and \code{\link{t50_bluck_coward}} for an example.
}
