% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breathtestfit_broom.R
\name{tidy.breathtestfit}
\alias{tidy.breathtestfit}
\title{Broom-style tidying methods for breathtestfit}
\usage{
\method{tidy}{breathtestfit}(x, ...)
}
\arguments{
\item{x}{Object of class \code{breathttestfit}}

\item{...}{other parameters passed to methods}
}
\value{
A tibble/data frame with columns
\itemize{
  \item{patient_id}{Patient Id (character)}
  \item{group}{Treatment or patient group (character)}
  \item{m}{Fraction metabolized}
  \item{k}{Time constant (1/minutes)}
  \item{beta}{The so-called lag parameters, no dimension}
  \item{t50}{Emptying half time in minutes as calculated following Maes/Ghoos}
}
}
\description{
Methods to tidy the results of class \code{breathttestfit} 
as generated by  \code{nls_fit} or \code{nlme_fit}. Returns the fit
coefficients and the Maes/Ghoos t50; additional parameters should be 
extracted from the key/value data in breattestfit.
}
