% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breathtest_parameters.R
\name{tlag_bluck_coward}
\alias{tlag_bluck_coward}
\title{lag phase for bluck_coward self-correcting fit}
\usage{
tlag_bluck_coward(cf)
}
\arguments{
\item{cf}{named vector of coefficients; only \code{k} and \code{beta} are required.
Note that in this package,  \code{k} is measured in 1/min (e.g. 0.01/min),
while in the literature is is often quoted as 1/h (e.g. 0.6/h).}
}
\value{
lag phase in minutes (time t at which the maximum in the rate of change
of g(t) occurs)
}
\description{
this parameter is misnamed: it is not the lag, but the maximum 
of the PDR time series.
}
\seealso{
\code{\link{exp_beta}}, and \code{\link{t50_bluck_coward}} for an example.
}
